split app code into smaller composables
