+++ app/src/main/java/com/github/nisrulz/samplezentone/ui/screen/main/PlayStopButton.kt
@@ -6,16 +6,21 @@ import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import com.github.nisrulz.samplezentone.R
 
 
 @Composable
 internal fun PlayStopButton(
+    modifier: Modifier = Modifier,
     isPlaying: Boolean,
     onClick: () -> Unit,
 ) {
-    FloatingActionButton(onClick = onClick) {
+    FloatingActionButton(
+        modifier = modifier,
+        onClick = onClick
+    ) {
         if (isPlaying) {
             Icon(
                 painterResource(R.drawable.stop),
